third_party/blink/web_tests/netinfo/network-quality-expected.txt

Tests that all rtt and downlink is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS connection.effectiveType is values[count][3]
PASS connection.rtt is values[count][4]
PASS connection.downlink is values[count][5]
PASS successfullyParsed is true

TEST COMPLETE
